+++ equ-clt/src/main/java/equ/clt/handler/lsterminal/LSTerminalHandler.java
@@ -37,9 +37,9 @@ public class LSTerminalHandler extends TerminalHandler {
     }
 
     @Override
-    public Map<Integer, SendTransactionBatch> sendTransaction(List transactionInfoList) throws IOException {
+    public Map<Long, SendTransactionBatch> sendTransaction(List transactionInfoList) throws IOException {
 
-        Map<Integer, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
+        Map<Long, SendTransactionBatch> sendTransactionBatchMap = new HashMap<>();
 
         for(Object transactionInfo : transactionInfoList) {
             Exception exception = null;
